Refactor App component and extract logic into custom hooks and helper functions
- Moved state management and related functions from App component to useAppState hook.
- Created separate helper functions for Mixpanel request handling and storage operations.
- Introduced EventDetailsPanel component for better separation of concerns.
- Removed unused theme and storage logic from App component.
- Updated imports and adjusted component structure accordingly.
