+++ src/App.tsx
@@ -1,278 +1,26 @@
-import { useState, useCallback, useEffect } from 'react';
-import { AppState, MixpanelRequest, MixpanelRequestType } from './types';
-import { MIXPANEL_API_PATTERN } from './constants';
+import { useAppState } from './hooks/useAppState';
 import { useMixpanelListener } from './hooks/useMixpanelListener';
-import { Toolbar, Theme } from './components/Toolbar';
-import { EventList } from './components/EventList';
-import { PropertyTable } from './components/PropertyTable';
+import { Toolbar } from './components/Toolbar';
+import { EventDetailsPanel } from './components/EventDetailsPanel';
 import { HelpPanel } from './components/HelpPanel';
 
-const STORAGE_KEY = 'SAVED_URL_KEY';
-const THEME_STORAGE_KEY = 'SAVED_THEME_KEY';
-const THEME_CYCLE: Theme[] = ['light', 'dark', 'auto'];
-
-function applyTheme(theme: Theme) {
-  document.documentElement.setAttribute('data-theme', theme);
-}
-
-async function getSavedTheme(): Promise<Theme> {
-  try {
-    const result = await chrome.storage.sync.get(THEME_STORAGE_KEY);
-    const value = result[THEME_STORAGE_KEY];
-    return THEME_CYCLE.includes(value) ? value : 'auto';
-  } catch {
-    return 'auto';
-  }
-}
-
-async function saveTheme(theme: Theme): Promise<void> {
-  try {
-    await chrome.storage.sync.set({ [THEME_STORAGE_KEY]: theme });
-  } catch {
-    // storage unavailable outside extension context
-  }
-}
-
-const BASE64_REGEX =
-  /^([0-9a-zA-Z+/]{4})*(([0-9a-zA-Z+/]{2}==)|([0-9a-zA-Z+/]{3}=))?$/;
-
-const initialState: AppState = {
-  requests: {},
-  selectedKey: null,
-  count: 0,
-  customAPIHost: '',
-  omitMixpanelProperties: true,
-  isRecording: true,
-  isBatched: false,
-};
-
-function getRequestType(url: string): MixpanelRequestType {
-  if (/\/engage/i.test(url)) return 'engage';
-  if (/\/groups/i.test(url)) return 'groups';
-  return 'track';
-}
-
-function getUrlParams(request: chrome.devtools.network.Request): Record<string, string> {
-  return (request.request.queryString ?? []).reduce<Record<string, string>>(
-    (acc, { name, value }) => {
-      acc[name] = value;
-      return acc;
-    },
-    {}
-  );
-}
-
-function getPostDataParam(request: chrome.devtools.network.Request): string {
-  const params = request.request.postData?.params ?? [];
-  const dataParam = params.find((p) => p.name === 'data');
-  return dataParam?.value ?? '';
-}
-
-function decodeProperties(encoded: string): unknown {
-  const decoded = decodeURIComponent(encoded);
-  const str = BASE64_REGEX.test(decoded) ? atob(decoded) : decoded;
-  return JSON.parse(str);
-}
-
-async function getSavedCustomHost(): Promise<string> {
-  try {
-    const result = await chrome.storage.sync.get(STORAGE_KEY);
-    const value = result[STORAGE_KEY];
-    return typeof value === 'string' ? value : '';
-  } catch {
-    return '';
-  }
-}
-
-async function saveCustomHost(host: string): Promise<void> {
-  try {
-    await chrome.storage.sync.set({ [STORAGE_KEY]: host });
-  } catch {
-    // storage unavailable (e.g. during dev outside extension context)
-  }
-}
-
-function downloadJson(data: unknown, filename: string): void {
-  const url =
-    'data:text/json;charset=utf-8,' + encodeURIComponent(JSON.stringify(data, null, 2));
-  const a = document.createElement('a');
-  a.href = url;
-  a.download = filename;
-  document.body.appendChild(a);
-  a.click();
-  a.remove();
-}
-
-function EventDetailsPanel({
-  count,
-  requests,
-  selectedKey,
-  omitMixpanelProperties,
-  onSelect,
-}: {
-  count: number;
-  requests: AppState['requests'];
-  selectedKey: string | null;
-  omitMixpanelProperties: boolean;
-  onSelect: (key: string) => void;
-}) {
-  return (
-    <>
-      <div className="panel panel--events">
-        <h3 className="panel-header">Events ({count})</h3>
-        <div className="panel-body">
-          <EventList requests={requests} selectedKey={selectedKey} onSelect={onSelect} />
-        </div>
-      </div>
-      <div className="panel panel--properties">
-        <h3 className="panel-header">Properties</h3>
-        <PropertyTable
-          selectedKey={selectedKey}
-          requests={requests}
-          omitMixpanelProperties={omitMixpanelProperties}
-        />
-      </div>
-    </>
-  );
-}
-
 export default function App() {
-  const [state, setState] = useState<AppState>(initialState);
-  const [theme, setTheme] = useState<Theme>('auto');
-  const [showHelp, setShowHelp] = useState(false);
-
-  useEffect(() => {
-    getSavedCustomHost().then((host) => {
-      setState((s) => ({ ...s, customAPIHost: host }));
-    });
-    getSavedTheme().then((t) => {
-      setTheme(t);
-      applyTheme(t);
-    });
-  }, []);
-
-  const isRequestValid = useCallback(
-    (request: chrome.devtools.network.Request, customAPIHost: string): boolean => {
-      const url = request.request.url;
-      if (!url) return false;
-      if (MIXPANEL_API_PATTERN.test(url)) return true;
-      if (customAPIHost && url.includes(customAPIHost)) return true;
-      return false;
-    },
-    []
-  );
-
-  const handleRequest = useCallback(
-    (request: chrome.devtools.network.Request) => {
-      setState((prev) => {
-        if (!prev.isRecording || !isRequestValid(request, prev.customAPIHost)) {
-          return prev;
-        }
-
-        try {
-          const method = request.request.method;
-          const urlParams = getUrlParams(request);
-          const type = getRequestType(request.request.url);
-          let newState = { ...prev };
-
-          const addEntry = (parsed: MixpanelRequest['data']): AppState => {
-            const count = newState.count + 1;
-            const key = `event-${count}`;
-            const entry: MixpanelRequest = { ...urlParams, type, data: parsed };
-            return {
-              ...newState,
-              count,
-              requests: { ...newState.requests, [key]: entry },
-              selectedKey: newState.selectedKey ?? key,
-            };
-          };
-
-          const isValidItem = (item: unknown): boolean => {
-            if (!item || typeof item !== 'object') return false;
-            const obj = item as Record<string, unknown>;
-            if (type === 'track') return typeof obj['event'] === 'string';
-            // engage/groups: must have at least one known operation key or any non-meta key
-            return Object.keys(obj).some((k) => k.startsWith('$') || k !== 'token');
-          };
-
-          if (method === 'GET') {
-            const encoded = urlParams['data'] ?? '';
-            const parsed = decodeProperties(encoded) as MixpanelRequest['data'];
-            if (isValidItem(parsed)) {
-              newState = addEntry(parsed);
-            }
-          } else if (method === 'POST') {
-            const encoded = getPostDataParam(request);
-            const parsed = decodeProperties(encoded);
-
-            if (Array.isArray(parsed)) {
-              newState = { ...newState, isBatched: true };
-              for (const item of parsed) {
-                if (isValidItem(item)) {
-                  newState = addEntry(item as MixpanelRequest['data']);
-                }
-              }
-            } else if (isValidItem(parsed)) {
-              newState = addEntry(parsed as MixpanelRequest['data']);
-            }
-          }
-
-          return newState;
-        } catch (err) {
-          console.error('handleRequest error:', err);
-          return prev;
-        }
-      });
-    },
-    [isRequestValid]
-  );
+  const {
+    state,
+    theme,
+    showHelp,
+    handleRequest,
+    handleToggleRecording,
+    handleClearAll,
+    handleToggleProperties,
+    handleDownload,
+    handleCustomHostChange,
+    handleSelectEvent,
+    handleCycleTheme,
+    handleToggleHelp,
+  } = useAppState();
 
   useMixpanelListener(handleRequest);
-
-  const handleToggleRecording = useCallback(() => {
-    setState((s) => ({ ...s, isRecording: !s.isRecording }));
-  }, []);
-
-  const handleClearAll = useCallback(() => {
-    setState((s) => ({
-      ...s,
-      requests: {},
-      selectedKey: null,
-      count: 0,
-      isBatched: false,
-    }));
-  }, []);
-
-  const handleToggleProperties = useCallback(() => {
-    setState((s) => ({ ...s, omitMixpanelProperties: !s.omitMixpanelProperties }));
-  }, []);
-
-  const handleDownload = useCallback(() => {
-    downloadJson(state.requests, 'mixpanel-events.json');
-  }, [state.requests]);
-
-  const handleCustomHostChange = useCallback((value: string) => {
-    const host = value.trim().toLowerCase();
-    setState((s) => ({ ...s, customAPIHost: host }));
-    saveCustomHost(host);
-  }, []);
-
-  const handleSelectEvent = useCallback((key: string) => {
-    setState((s) => ({ ...s, selectedKey: key }));
-  }, []);
-
-  const handleCycleTheme = useCallback(() => {
-    setTheme((prev) => {
-      const next = THEME_CYCLE[(THEME_CYCLE.indexOf(prev) + 1) % THEME_CYCLE.length];
-      applyTheme(next);
-      saveTheme(next);
-      return next;
-    });
-  }, []);
-
-  const handleToggleHelp = useCallback(() => {
-    setShowHelp((v) => !v);
-  }, []);
 
   return (
     <div className="app">
