+++ src/helpers/theme.helpers.ts
@@ -1,0 +1,5 @@
+import { Theme } from '../components/Toolbar';
+
+export function applyTheme(theme: Theme): void {
+  document.documentElement.setAttribute('data-theme', theme);
+}
